Update TS feature to use new single-file approach
