+++ package.json
@@ -9,7 +9,7 @@
   },
   "dependencies": {
     "commander": "^8.3.0",
-    "temporalio": "1.0.0-rc.1",
+    "temporalio": "1.0.1",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
